#612 - Split AnnotatorState into multiple parts
- Moved default methods  to AnnotatorViewState interface.
- Moved default methods  to AnnotatorDocumentNavigation interface.
